Funcionalidades casi completas , tabajando en las sesiones

+++ src/main/java/com/ipc2/proyectofinalservlet/model/CargarDatos/Comision.java
@@ -2,6 +2,8 @@ package com.ipc2.proyectofinalservlet.model.CargarDatos;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @ToString
@@ -9,5 +11,5 @@ import lombok.*;
 @AllArgsConstructor
 public class Comision {
     private int codigo;
-    private int cantidad;
+    private BigDecimal cantidad;
 }
